Avoid inclusion of Vc/Vc in RandomRing
Inclusiong of Vc/Vc is argued to cause
gcc10 to take lot's of memory leading to crashes
in the CI.

In this particular case, I think we can avoid
the inclusion og Vc/Vc in this header and leave
it it to the source files to do so.
(But this might not always be possible);
