+++ Detectors/FIT/FT0/simulation/include/FT0Simulation/Digitizer.h
@@ -37,7 +37,7 @@ class Digitizer
 {
  private:
   using DP = DigitizationConstants;
-  typedef math_utils::RandomRing<float_v::size() * DP::NOISE_RANDOM_RING_SIZE> NoiseRandomRingType;
+  typedef math_utils::RandomRing</*float_v::size()*/ 4 * DP::NOISE_RANDOM_RING_SIZE> NoiseRandomRingType;
 
  public:
   Digitizer(Int_t mode = 0) : mMode(mode), mRndGaus(NoiseRandomRingType::RandomType::Gaus), mNumNoiseSamples(), mNoiseSamples(), mSincTable(), mSignalTable(), mSignalCache() { initParameters(); }
@@ -109,13 +109,15 @@ class Digitizer
     return mSignalTable[index] + rem * (mSignalTable[index + 1] - mSignalTable[index]);
   }
 
-  inline Vc::float_v signalFormVc(Vc::float_v x) const
+  template <typename VcType>
+  inline VcType signalFormVc(VcType x) const
   { // table lookup for the signal shape (SIMD version)
+    // implemented as template function, so that we don't need to include <Vc/Vc> here
     auto const y = x / DigitizationParameters::Instance().mBunchWidth * DP::SIGNAL_TABLE_SIZE;
-    Vc::float_v::IndexType const index = Vc::floor(y);
+    typename VcType::IndexType const index = floor(y);
     auto const rem = y - index;
-    Vc::float_v val(0);
-    for (size_t i = 0; i < float_v::size(); ++i) {
+    VcType val(0);
+    for (size_t i = 0; i < VcType::size(); ++i) {
       if (y[i] < 0.0f) {
         continue;
       }
@@ -159,40 +161,6 @@ class Digitizer
   ClassDefNV(Digitizer, 2);
 };
 
-// signal shape function
-template <typename Float>
-Float signalForm_i(Float x)
-{
-  using namespace std;
-  Float const a = -0.45458;
-  Float const b = -0.83344945;
-  return x > Float(0) ? -(exp(b * x) - exp(a * x)) / Float(7.8446501) : Float(0);
-  //return -(exp(-0.83344945 * x) - exp(-0.45458 * x)) * (x >= 0) / 7.8446501; // Maximum should be 7.0/250 mV
-};
-
-// integrated signal shape function
-inline float signalForm_integral(float x)
-{
-  using namespace std;
-  double const a = -0.45458;
-  double const b = -0.83344945;
-  if (x < 0) {
-    x = 0;
-  }
-  return -(exp(b * x) / b - exp(a * x) / a) / 7.8446501;
-};
-
-// SIMD version of the integrated signal shape function
-inline Vc::float_v signalForm_integralVc(Vc::float_v x)
-{
-  auto const mask = (x >= 0.0f);
-  Vc::float_v arg(0);
-  arg.assign(x, mask); // branchless if
-  Vc::float_v const a(-0.45458f);
-  Vc::float_v const b(-0.83344945f);
-  Vc::float_v result = -(Vc::exp(b * arg) / b - Vc::exp(a * arg) / a) / 7.8446501f;
-  return result;
-};
 } // namespace ft0
 } // namespace o2
